Introduce new package core.routing
