+++ javalin/src/main/java/io/javalin/http/util/RedirectToLowercasePathPlugin.kt
@@ -7,7 +7,7 @@
 package io.javalin.http.util
 
 import io.javalin.Javalin
-import io.javalin.core.PathSegment
+import io.javalin.core.routing.PathSegment
 import io.javalin.core.createPathParser
 import io.javalin.core.plugin.Plugin
 import io.javalin.core.plugin.PluginLifecycleInit
